pyproject.toml
[project]
name = "retrival-augment-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
]
